update: switch rootpage timeline to versions
