+++ src/pages/index.tsx
@@ -1,264 +1,218 @@
-import ActionIcons from "@/components/timeline/actionIcons";
-import { Post } from "@/utils/types";
+import ConfirmationModal from "@/components/common/confirmationModal";
+import { fetchVersionsData } from "@/features/pokemons";
+import { RootState } from "@/redux/store";
+import { Pokemon, Version } from "@/utils/types";
+import { loginPath, timelinePath } from "@/utils/urls/client";
 import { apiLocalhost } from "@/utils/urls/server";
 import {
-  Alert,
+  Box,
+  Button,
   Card,
-  Divider,
-  Fab,
+  CardContent,
+  Modal,
   TextareaAutosize,
-  TextField,
+  Typography,
 } from "@mui/material";
+import { GetStaticProps } from "next";
 import Image from "next/image";
-import { useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
-import { useDispatch, useSelector } from "react-redux";
-import styles from "../components/timeline/timeline.module.scss";
-import { RootState } from "@/redux/store";
-import { updatePosts, updateSelectedPost } from "@/redux/reducers/posts";
-import ConfirmationModal from "@/components/common/confirmationModal";
-import Comments from "@/components/timeline/comments";
-import InfiniteScroll from "react-infinite-scroll-component";
-import ScrollTop from "@/components/timeline/scrollTop";
-import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
-import Head from "next/head";
+import Router from "next/router";
+import React, { useState } from "react";
+import { FieldValues, useForm } from "react-hook-form";
+import { useSelector } from "react-redux";
+import styles from "../components/versions.module.scss";
 
-const TimeLine = () => {
-  const dispatch = useDispatch();
-  const { currentUser, isLoggedIn } = useSelector(
+export const getStaticProps: GetStaticProps = async () => {
+  const versions = await fetchVersionsData();
+  return {
+    props: {
+      versions,
+    },
+  };
+};
+type Props = {
+  versions: Version[];
+};
+const Index = ({ versions }: Props) => {
+  const { isLoggedIn, currentUser } = useSelector(
     (state: RootState) => state.authReducer
   );
-  const { posts, selectedPost } = useSelector(
-    (state: RootState) => state.postsReducer
-  );
-
-  useEffect(() => {
-    (async () => {
-      const { posts } = await apiLocalhost
-        .get("/posts")
-        .then((res) => res.data);
-      dispatch(updatePosts(posts));
-    })();
-  }, []);
-
-  const handlePostSelect = (post: Post) => {
-    if (selectedPost.post?.id === post.id) return;
-    dispatch(updateSelectedPost({ post, editing: false }));
-  };
-  const [commentingPost, setCommentingPost] = useState<Post | undefined>(
+  const [modalOpen, setModalOpen] = useState(false);
+  const [selectedVersion, setSelectedVersion] = useState<Version | undefined>(
     undefined
   );
-
-  const handleClickMessageIcon = () => {
-    setCommentingPost(selectedPost.post);
-  };
-  const handleClickEditIcon = () => {
-    if (selectedPost.post === undefined) return;
-    dispatch(updateSelectedPost({ ...selectedPost, editing: true }));
+  const handleOpenModal = (version: Version) => {
+    setModalOpen(true);
+    setSelectedVersion(version);
   };
 
-  const { register, handleSubmit } = useForm();
-
-  type onSubmitProp = {
-    title: string;
-    content: string;
+  const handleCloseModal = () => {
+    setModalOpen(false);
+    setSelectedVersion(undefined);
+    setSelectedPokemon(undefined);
   };
-  const onSubmit = async (data: any) => {
-    const { title, content } = data as onSubmitProp;
-    // API叩く
-    const { post } = await apiLocalhost
-      .patch(`/posts/${selectedPost.post?.id}`, {
-        posts: { title, content },
-      })
-      .then((res) => res.data);
-    const updatedPosts = posts?.map((tgtPost) => {
-      if (tgtPost.id === post.id) {
-        return post;
-      } else {
-        return tgtPost;
-      }
+
+  const [selectedPokemon, setSelectedPokemon] = useState<Pokemon | undefined>(
+    undefined
+  );
+  const handleSelectPokemon = (pokemon: Pokemon) => {
+    // ポケモンを選択した場合も、ログインしていない場合はログインモーダルを表示する
+    checkLogin();
+    setSelectedPokemon(pokemon);
+  };
+
+  const { register, handleSubmit, reset } = useForm<FieldValues>();
+  const onSubmit = async (formData: any) => {
+    checkLogin();
+    const { title, content } = formData as { title: string; content: string };
+    const { status } = await apiLocalhost.post("/posts", {
+      posts: {
+        title,
+        content,
+        pokemon_name: selectedPokemon!.name,
+        pokemon_image: selectedPokemon!.image,
+        version_name: selectedVersion!.name,
+        generation_name: selectedVersion!.data.generation.name,
+      },
     });
-    dispatch(updatePosts(updatedPosts));
-    dispatch(
-      updateSelectedPost({
-        ...selectedPost,
-        editing: false,
-        post: post,
-      })
-    );
-  };
-  const [open, setOpen] = useState(false);
-  const handleDeleteModalOpen = () => {
-    setOpen(true);
-  };
-  const handleDeleteModalClose = () => {
-    setOpen(false);
-  };
-  const deletePost = async () => {
-    const postId = selectedPost.post?.id;
-    await apiLocalhost.delete(`/posts/${postId}`);
-    const updatedPosts = posts?.filter((post) => post.id !== postId);
-    dispatch(updatePosts(updatedPosts));
-    dispatch(updateSelectedPost({ post: undefined, editing: false }));
-    handleDeleteModalClose();
-  };
-  const actionIconsFuncs = {
-    handleClickMessageIcon,
-    handleClickEditIcon,
-    handleClickDeleteIcon: handleDeleteModalOpen,
-  };
-  const [hasMore, setHasMore] = useState(false);
-  const [page, setPage] = useState(1);
-  const loadComments = async () => {
-    const morePosts = await apiLocalhost
-      .get(`/posts?page=${page}`)
-      .then((res) => res.data.posts);
-    if (morePosts.length < 10) {
-      setHasMore(false);
+    if (status === 200) {
+      reset();
+      Router.push(timelinePath);
     }
-    setPage(page + 1);
-    dispatch(updatePosts([...posts!, ...morePosts]));
   };
-  useEffect(() => {
-    const updatedSelectedPost = posts?.find(
-      (post) => post.id === selectedPost.post?.id
-    );
-    dispatch(
-      updateSelectedPost({
-        post: updatedSelectedPost,
-        editing: selectedPost.editing,
-      })
-    );
+  const checkLogin = () => {
+    if (isLoggedIn && currentUser) return;
 
-    if (posts && posts.length < 11) {
-      setHasMore(true);
-    }
-  }, [posts]);
+    setOpenLoginModal(true);
+  };
+  const [openLoginModal, setOpenLoginModal] = useState(false);
+  const handleCloseLoginModal = () => {
+    setOpenLoginModal(false);
+  };
+  const handleJumpLoginPage = () => {
+    Router.push(loginPath);
+  };
 
   return (
-    <>
-      <Head>
-        <link rel="icon" href="favicon.ico" />
-      </Head>
-      <div style={{ width: "90%", margin: "0 auto" }}>
-        <ConfirmationModal
-          handleClose={handleDeleteModalClose}
-          confirmationTxt="本当にこの投稿を削除しますか？"
-          execFunc={deletePost}
-          open={open}
-        />
-        {posts && posts.length > 0 ? (
-          <div className={styles.cardsWrapper}>
-            <InfiniteScroll
-              dataLength={posts.length}
-              next={loadComments}
-              hasMore={hasMore}
-              endMessage={
-                <h2 style={{ textAlign: "center" }}>
-                  これ以上投稿はありません
-                </h2>
-              }
-              loader={<h4 style={{ textAlign: "center" }}>読み込み中</h4>}
+    <div style={{ width: "90%", margin: "0 auto" }}>
+      <ConfirmationModal
+        handleClose={handleCloseLoginModal}
+        confirmationTxt={"投稿にはログインが必要です。ログインしますか？"}
+        execFunc={handleJumpLoginPage}
+        open={openLoginModal}
+      />
+      <div>
+        {versions?.map((version, index) => {
+          return (
+            <Card
+              key={index}
+              style={{ margin: "1em" }}
+              className={styles.versionCard}
+              onClick={() => handleOpenModal(version)}
             >
-              {posts.map((post, index) => {
-                return (
-                  <Card
-                    id="back-to-top-anchor"
-                    key={index}
-                    className={
-                      post.id === selectedPost.post?.id
-                        ? styles.selectedPost
-                        : styles.postCard
-                    }
-                    onClick={() => handlePostSelect(post)}
-                  >
-                    {post.id === selectedPost.post?.id && (
-                      <>
-                        <ActionIcons
-                          {...actionIconsFuncs}
-                          createdByCurrentUser={
-                            post.user_id === currentUser?.id
-                          }
-                          commentsLength={post.comments_length}
-                        />
-                      </>
-                    )}
-                    <div className={styles.pokemonImageWrapper}>
-                      <Image
-                        className={styles.pokemonImage}
-                        src={post.pokemon_image}
-                        alt={post.pokemon_name}
-                        width={100}
-                        height={100}
-                      />
-                    </div>
-                    <div className={styles.postHeader}>
-                      <h1>{post.pokemon_name}</h1>
-                      <h2 className={styles.test}>{post.version_name}</h2>
-                      <h4>{post.generation_name}</h4>
-                    </div>
-
-                    <Alert
-                      className={styles.userContainer}
-                      icon={false}
-                      variant="outlined"
-                      severity="info"
-                    >
-                      <div>
+              <CardContent>
+                <Typography variant="h5" className={styles.versionName}>
+                  {version.name}
+                </Typography>
+                <p className={styles.generationName}>
+                  {version.data.generation.name}
+                </p>
+                <div className={styles.pokemonsWrapper}>
+                  {version.data.pokemons.map((pokemon, index) => {
+                    return (
+                      <div key={index} className={styles.pokemonWrapper}>
                         <Image
-                          src={post.user_img}
-                          alt={post.user_name}
-                          width={100}
-                          height={100}
+                          height={40}
+                          width={40}
+                          src={pokemon.image}
+                          alt={pokemon.name}
                         />
-                        <span>{post.user_name}</span>
+                        <p>{pokemon.name}</p>
                       </div>
-                    </Alert>
-                    <div className={styles.postContainer}>
-                      {selectedPost.post?.id === post.id &&
-                      selectedPost.editing ? (
-                        <div className={styles.postFormWrapper}>
-                          <form onSubmit={handleSubmit(onSubmit)}>
-                            <button>編集を完了する</button>
-                            <TextField
-                              id="filled-basic"
-                              label="タイトル"
-                              variant="filled"
-                              type="text"
-                              defaultValue={selectedPost.post?.title}
-                              {...register("title", { required: true })}
-                            />
-                            <TextareaAutosize
-                              defaultValue={selectedPost.post?.content}
-                              {...register("content", { required: true })}
-                            />
-                          </form>
-                        </div>
-                      ) : (
-                        <>
-                          <h2>{post.title}</h2>
-                          <p>{post.content}</p>
-                        </>
-                      )}
-                    </div>
-                    <Divider />
-                    {commentingPost?.id === post.id && <Comments />}
-                  </Card>
-                );
-              })}
-            </InfiniteScroll>
-          </div>
-        ) : (
-          <div className={styles.noPostsMsg}>投稿がありません</div>
-        )}
+                    );
+                  })}
+                </div>
+                <p>
+                  {version.data.generation?.regions.map((region, index) => {
+                    return (
+                      <>
+                        {version.data.generation?.regions.length > 1 &&
+                          index !== 0 && <span>&nbsp;&nbsp;</span>}
+                        <span key={index} className={styles.versionName}>
+                          {region?.name}
+                        </span>
+                      </>
+                    );
+                  })}
+                </p>
+              </CardContent>
+            </Card>
+          );
+        })}
       </div>
-      <ScrollTop>
-        <Fab size="small" aria-label="scroll back to top">
-          <KeyboardArrowUpIcon />
-        </Fab>
-      </ScrollTop>
-    </>
+
+      <Modal
+        open={modalOpen}
+        onClose={handleCloseModal}
+        className={styles.createPostModal}
+      >
+        <>
+          <Button onClick={handleCloseModal}>Close</Button>
+          {selectedVersion && (
+            <>
+              <Typography variant="h5">{selectedVersion.name}</Typography>
+              <p>{selectedVersion.data.generation.name}</p>
+              <div className={styles.starterPokemonsWrapper}>
+                {selectedVersion.data.pokemons.map((pokemon, index) => {
+                  return (
+                    <Box
+                      key={index}
+                      onClick={() => handleSelectPokemon(pokemon)}
+                      className={styles.starterPokemon}
+                    >
+                      <Image
+                        width={120}
+                        height={120}
+                        src={pokemon.image}
+                        alt={pokemon.name}
+                        className={
+                          pokemon === selectedPokemon
+                            ? styles.starterPokemonImg
+                            : styles.starterPokemonImgSelected
+                        }
+                      />
+                      <p>{pokemon.name}</p>
+                    </Box>
+                  );
+                })}
+              </div>
+            </>
+          )}
+          <form
+            onSubmit={handleSubmit(onSubmit)}
+            className={styles.createPostForm}
+          >
+            <input
+              onClick={checkLogin}
+              placeholder="Title"
+              {...register("title", { required: true })}
+              className={styles.createPostFormTitle}
+            />
+            <TextareaAutosize
+              onClick={checkLogin}
+              placeholder="Description"
+              {...register("content", { required: true })}
+              className={styles.createPostFormContent}
+            />
+            {isLoggedIn && selectedPokemon && (
+              <Button type="submit" className={styles.createPostFormSubmitBtn}>
+                投稿する
+              </Button>
+            )}
+          </form>
+        </>
+      </Modal>
+    </div>
   );
 };
 
-export default TimeLine;
+export default Index;
