+++ src/utils/urls/client.ts
@@ -1,4 +1,4 @@
 export const singUpPath = "/auth/signup";
 export const loginPath = "/auth/login";
-export const versionsPath = "/versions";
-export const timelinePath = "/";
+export const timelinePath = "/timeline";
+export const versionsPath = "/";
